chore(frontend): consolidate AvatarDisplay to directory component (#281)
The flat AvatarDisplay.tsx now re-exports from the canonical
AvatarDisplay/index.tsx directory component, preventing future
divergence between the two implementations.
